first settings implementation + code cleanup
i have started working on settings for the client. they aren't perfect
but they should do for now :)
i have also implemented more data binding stuff which makes it easier
and cleaner. i need to implement a custom event for when the backend is
setting up and stuff but thats for later